# Src/meetDoc/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MeetDoc",
        "description": "MeetDoc APIs Documentation using Swagger",
        "contact": {
            "email": "mohamed.ahmed.aly.d@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/email/reset-password": {
            "post": {
                "tags": [
                    " Email - admin reset password -"
                ],
                "summary": "admin send eamil to reset password",
                "operationId": "9d0da40a013f7f97d472641763c46f1e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/email/confirm-password-change": {
            "post": {
                "tags": [
                    " Email - admin reset password -"
                ],
                "summary": "confirm password change",
                "operationId": "79038d21b3e9539d8a084ffadde14b65",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirmPassword": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/appointments/": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "get all apointments",
                "operationId": "374e3b5733c5a54710159c9885776b4b",
                "responses": {
                    "200": {
                        "description": "all appointments",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cancel-appointment/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "canceling an appointment",
                "operationId": "0f69d42207abba6a21ef879224b5fd5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of appointment",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "all appointments",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/appointments/info/": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "get patient name and service type",
                "description": "get patient name and service type by serviceType and date (mobile)",
                "operationId": "8f60cf708d23067d1d9d3c05da86328e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string"
                                    },
                                    "serviceType": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/upload": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "uploads an image or pdf",
                "description": "upload medical documents related to th epatient",
                "operationId": "uploadFile",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "documents[]": {
                                        "description": "file to upload",
                                        "type": "string",
                                        "format": "base64"
                                    },
                                    "patientId": {
                                        "description": "patient id",
                                        "type": "integer"
                                    },
                                    "createdBy": {
                                        "description": "creator id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "All Patient Symptoms returned successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}